+++ src/data/collections/array/d1/methods/mod.rs
@@ -1,0 +1,12 @@
+// devela::data::collections::array::d1::methods
+//
+//! 1-dimensional array methods
+//
+
+mod common;
+
+mod bare;
+
+#[cfg(feature = "alloc")]
+#[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "alloc")))]
+mod boxed;